created
A bank accepts fixed deposits for one year or more and the policy it adopts on interest is as follows (1) If a deposit is less than Rs 2000 and for 2 years or more years, the interest rate is 5 percent compound annually. (2) If a deposit is Rs 2000 or more but less than 6000 and for 2 or more years , the interest rate is 7 percent compounded annually. (3) On all deposit for 5 years or more, interest is 10 percent compounded annually (4) On all other deposits not covered by the above conditions, the interest is 3 percent compounded annually. Give the amount deposited and the number of years, write a program to calculate the money in the customers' account at the end of the specified time.